+++ CHANGELOG.md
@@ -63,6 +63,12 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
     This decorator is automatically added during the build process.
   * Optimized creation of metaobjects.
 
+## v4.0.0-beta.152 (2024-11-11)
+
+#### :rocket: New Feature
+
+* Added `trackContentSwipes` - a flag to prevent unexpected closure of a component instance `bBottomSlide`
+
 ## v4.0.0-beta.151 (2024-11-06)
 
 #### :bug: Bug Fix